Move existing ui elements to ui.simple package
