+++ indigo/indigo-extras/src/main/scala/indigoextras/ui/simple/InputField.scala
@@ -1,4 +1,4 @@
-package indigoextras.ui
+package indigoextras.ui.simple
 
 import indigo.shared.BoundaryLocator
 import indigo.shared.Context
